+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/ImageUtils.kt
@@ -18,9 +18,14 @@ package io.getstream.chat.android.compose.ui.util
 
 import android.content.Context
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.graphics.Brush
@@ -29,11 +34,13 @@ import androidx.compose.ui.graphics.FilterQuality
 import androidx.compose.ui.graphics.drawscope.DrawScope
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.LayoutDirection
 import coil.compose.AsyncImagePainter
 import coil.compose.rememberAsyncImagePainter
 import coil.request.ImageRequest
+import coil.size.Size
 import com.skydoves.landscapist.ImageOptions
 import com.skydoves.landscapist.coil.CoilImage
 import com.skydoves.landscapist.coil.CoilImageState
@@ -82,6 +89,77 @@ public fun Modifier.mirrorRtl(layoutDirection: LayoutDirection): Modifier {
         scaleX = if (layoutDirection == LayoutDirection.Ltr) 1f else -1f,
         scaleY = 1f,
     )
+}
+
+/**
+ * Displays an image asynchronously using `Coil` and the [LocalStreamImageLoader].
+ * It transforms the image URL and provides headers before loading the image.
+ *
+ * @param data The data to load the image from. Can be a URL, URI, resource ID, etc.
+ * @param modifier Modifier for styling.
+ * @param contentScale The scale to be used for the content.
+ * @param content A composable function that defines the content to be displayed based on the image loading state.
+ *
+ * @see ImageAssetTransformer
+ * @see ImageHeadersProvider
+ */
+@Composable
+internal fun StreamAsyncImage(
+    data: Any?,
+    modifier: Modifier = Modifier,
+    contentScale: ContentScale = ContentScale.Fit,
+    content: @Composable BoxScope.(state: AsyncImagePainter.State) -> Unit,
+) {
+    StreamAsyncImage(
+        imageRequest = ImageRequest.Builder(LocalContext.current)
+            .data(data)
+            .build(),
+        modifier = modifier,
+        contentScale = contentScale,
+        content = content,
+    )
+}
+
+/**
+ * Displays an image asynchronously using `Coil` and the [LocalStreamImageLoader].
+ * It transforms the image URL and provides headers before loading the image.
+ *
+ * @see ImageAssetTransformer
+ * @see ImageHeadersProvider
+ *
+ * @param imageRequest The request to load the image.
+ * @param modifier Modifier for styling.
+ * @param contentScale The scale to be used for the content.
+ * @param content A composable function that defines the content to be displayed based on the image loading state.
+ */
+@Composable
+internal fun StreamAsyncImage(
+    imageRequest: ImageRequest,
+    modifier: Modifier = Modifier,
+    contentScale: ContentScale = ContentScale.Fit,
+    content: @Composable BoxScope.(state: AsyncImagePainter.State) -> Unit,
+) {
+    var imageSize by remember { mutableStateOf(Size.ORIGINAL) }
+    Box(
+        modifier = modifier.onSizeChanged { size -> imageSize = Size(size.width, size.height) },
+    ) {
+        if (imageSize == Size.ORIGINAL) {
+            content(AsyncImagePainter.State.Empty)
+        } else {
+            val context = LocalContext.current
+            val imageAssetTransformer = ChatTheme.streamImageAssetTransformer
+            val imageHeaderProvider = ChatTheme.streamImageHeadersProvider
+            val asyncImagePainter = rememberAsyncImagePainter(
+                model = imageRequest
+                    .convertUrl(context, imageAssetTransformer)
+                    .provideHeaders(context, imageHeaderProvider)
+                    .size(context, imageSize),
+                imageLoader = LocalStreamImageLoader.current,
+                contentScale = contentScale,
+            )
+            content(asyncImagePainter.state)
+        }
+    }
 }
 
 /**
@@ -350,6 +428,15 @@ private fun ImageRequest.provideHeaders(
                 .forEach { addHeader(it.key, it.value) }
         }
         .build()
+
+/**
+ * Set the [Size] as a new build of the [ImageRequest].
+ */
+private fun ImageRequest.size(context: Context, size: Size): ImageRequest = run {
+    newBuilder(context)
+        .size(size)
+        .build()
+}
 
 /**
  * Used to change a parameter set on Coil requests in order